[Bug Fix] Logarithmic Average function added
Symptoms: Generation labels on Uckeleoids go to NaN shortly after the
60th generation

Reason: The generation was being computed off of the ancestor count and
ancestor count was being stored in a long. A 60th generation Uckleoid
has approximately 2^60 (non-unique) ancestors, which stresses the limits
of a 64 bit long.

Solution: By storing only the generation itself, we can take the average
of two numbers on a log scale. This avoids storage problems of actually
keeping the full ancestor count in memory.