Implemented a working second brain type, Random
Random brains generate random output. This is useful only as far as it
is a practical test of structure for running multiple neurological models
in the same environment (and also implicitly, the future ability to
configure neurological models through a setup script). Also, this can
be useful to establish baseline performance of different 'blind'
probabilistic creatures to compare what advantages the NN models are able
to generate.