Reduced the level of association between World and Creature objects.
This was done to make it possible to create a Creature to use in a genetic
algorithm without connecting it to a world, and potentially running it
in multiple worlds or adding it to a world later.

This change breaks Census and ActionProfile code, as these features don't
make cohesive sense in the shifted programming paradigm. I want to move
forward with decoupled World and Creature objects for generalized
genetic algorithm based simulations before restoring these features.