Vivarium-web dropped
I've decided to drop vivarium web due to extensive contortions caused by
the lack of support for any serious reflection in GWT. My needs to run the
visualizer are minor, and while a plugin-free web based visualizer is a
nice to have feature, I would rather focus on being able to iterate quickly
on genetic algorithm or neurology code.

I decided not to drop vivarium-web unless I could build a reasonable
replacement first, and that is the culmination of this branch. This branch
introduced vivarium-swing, a Java application, swing based, animated
visualizer. Early versions of this were visually identical to vivarium-web,
but tend to run much faster. Moving forward, there are a lot of visualizer
improvements that can be made, and vivarium-swing will eventually far
exceed vivarium-web in functionality.