[47] Added UUIDs to all objects
This commit removes the interface MapSerializer and replaces it with the
abstract class VivariumObject. This is partly a naming change, but it also
represents a change in assumption that all durable objects in Vivarium
can share properties (such as having a UUID)

This change includes a rewrite of the UUID for compatibility with GWT
and GWT Streamable.