Removed unused code (HierarchicalListParser)
The HierarchicalListParser was used when I was trying to keep my JSON in
a certain limited flat structure due restrictions to support GWT.  Since
I have moved my serialization code so that the web code does not rely on
the serializer, I've been free to use much more robust reflective methods.

This code was a terrible hack, and I'm glad to see it go.