[2] Simulation as a top level VivariumObject has been dropped.
I've found the Simulation object to entirely insufficient to describe
job behavior for scripts, redundant with the RunSimulation script which
doesn't even require a Simulation (it can work with just a World), and
finally Simulation doesn't match other VivariumObjects which tend to be
very data oriented. Simulation has been removed, to allow its functionality
to be focused into Scripts and Jobs.