[49] Refactor to allow better script isolation
This refactor makes it easier to seperate a script's functionality
from file IO (useful for benchmarks or workers who do not want to use
file IO).

This only adds a new functionality only run mode to RunSimulation, although
many other scripts could be updated to work like this as a later task.