Added Breathing + Janky transition between interpolated & not
The current transition between interpolated & not can cause a weird 'flashback'
This is still better than the original naive approach, because we won't spend
doing world copies for snapshots when the simulation is running too fast
to use interpolation. We should detect the transition better and buffer two
snapshots before switching the animation mode.