Only simulate creatures in the center
Creatures on the edge of the world will cause out of bounds exceptions
on arrays by trying to detect off the edge of the world. This is
historically fixed by keeping the edges of the world filled with walls.

With the addition of the interactive visualizer, it's now possible to
let a user delete an edge wall, which invariably crashes the simulation
as soon as a creature moves into this space. As a short term safeguard,
that should also marginally improve performance, creatures that move
into the edge of the world will simply stop getting updated.

In the future, it might be nice to allow creatures to move to the edge
of a World in order to migrate to other Worlds, and creatures will be
be able to see more than a single square in front of themselves, so
this modification might become irrelevant.